moves syncOrcidProfile to data-provider
